// frontend/src/shared/lib/api/model/datasetReadinessResponse.ts
/**
 * Generated by orval. Do not edit manually.
 * daikon-studio
 */
import type { DatasetReadinessResponsePartitionCounts } from './datasetReadinessResponsePartitionCounts';
import type { TargetClassBalanceResponse } from './targetClassBalanceResponse';

export interface DatasetReadinessResponse {
  row_count: number;
  partition_counts: DatasetReadinessResponsePartitionCounts;
  class_balance: TargetClassBalanceResponse[];
  warnings: string[];
}